Build 116: JOHN v3, HTMLParser
AnyResult: generic and/or type-safe interface to access the output of a JOHN plugin via a known interface and conditional values.
Subscript: advanced subscript enabling placeholders, ranges and catch-all.
Parameters: before-execution verification of the required parameters, definition of default values, and setting of authorization credentials as internally defined variables.
IOMarkup+XML now integrates an HTMLParser that is a drop-in replacement for the FoundationXML.XMLParser already used for XML; IOMarkup now supports case-insensitive access of attributes and tag names, as is the case with HTML hierarchies.
Redirection enables individual stages to choose if and how many redirects are allowed, making it possible to retreive information from HTTP headers of a 302 redirect response.
Cookies are now supported, with a special stage parameter to manipulate them (either setting a value or deleting it), inheritance policy following RFC 6265, and the ability to yield them instead of the request headers or body.
DebugDelegate is a protocol that allows for the inspection of the execution state of the plugin and LoggerDelegate is a convenience implementation of it that logs the execution state to Standard Output or any handler conforming to the Logger protocol provided by the swift-log library.
Identified requests, Defered requests and the execution Group enables the sharing of responses between any number of requests execution on said group, while the deferral of requests coalesce cleanup operations at the very end of the execution group.
—
Fixed refresh token being deleted when refreshing the access token.